// src/types/operationTaskCenter.ts
export enum OperationTaskEnum {
    generateOperationCreep = 'generateOperationCreep',
    needHarvester = 'needHarvester',
    needBuilder = 'needBuilder',
    needUpgrader = 'needUpgrader',
    needRepairer = 'needRepairer',
    transformIdleOperationCreepToUpgrader = 'transformIdleOperationCreepToUpgrader',
}